+++ NEWS.md
@@ -1,6 +1,7 @@
 ## Change in v0.5.1
 
 - Add `normalize` to `textmodel_doc2vec()` and pass it to `as.matrix()`.
+- Add `weight` to `textmodel_doc2vec()` to adjust the salience of words in the resulting document vectors.
 
 ## Changes in v0.5.0