+++ NEWS.md
@@ -1,6 +1,6 @@
 ## Change in v0.5.1
 
-- Add `normalize` to `textmodel_doc2vec()` and pass it to `as.matrix()`.
+- Normalize document vectors by the lengths of documents when `normalize = TRUE`.
 - Add `weights` to `textmodel_doc2vec()` to adjust the salience of words in the document vectors.
 - Add `include_data` to `textmodel_word2vec()` to save the original tokens object.